remove icc warnings
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@8695 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ tests/thin_client.c
@@ -100,7 +100,7 @@ do_connection(char * remote_host, int port)
     } else {
 	memcpy((char*)&sin.sin_addr, host_addr->h_addr, host_addr->h_length);
     }
-    sin.sin_port = htons(port);
+    sin.sin_port = (unsigned short) htons(port);
     sin.sin_family = AF_INET;
     if (connect(conn, (struct sockaddr *) &sin,
 		sizeof sin) == -1) {
